refactor(editor): Organize constants into separate files (no-changelog)
