+++ packages/frontend/editor-ui/src/constants/actions.ts
@@ -1,0 +1,13 @@
+export const enum WORKFLOW_MENU_ACTIONS {
+	DUPLICATE = 'duplicate',
+	DOWNLOAD = 'download',
+	IMPORT_FROM_URL = 'import-from-url',
+	IMPORT_FROM_FILE = 'import-from-file',
+	PUSH = 'push',
+	SETTINGS = 'settings',
+	DELETE = 'delete',
+	ARCHIVE = 'archive',
+	UNARCHIVE = 'unarchive',
+	RENAME = 'rename',
+	CHANGE_OWNER = 'change-owner',
+}
