+++ packages/frontend/editor-ui/src/constants/breakpoints.ts
@@ -1,0 +1,4 @@
+export const BREAKPOINT_SM = 768;
+export const BREAKPOINT_MD = 992;
+export const BREAKPOINT_LG = 1200;
+export const BREAKPOINT_XL = 1920;
