+++ packages/frontend/editor-ui/src/constants/emails.ts
@@ -1,0 +1,3 @@
+export const N8N_SALES_EMAIL = 'sales@n8n.io';
+
+export const N8N_CONTACT_EMAIL = 'contact@n8n.io';
