+++ packages/frontend/editor-ui/src/constants/injectionKeys.ts
@@ -1,0 +1,20 @@
+import type {
+	CanvasInjectionData,
+	CanvasNodeHandleInjectionData,
+	CanvasNodeInjectionData,
+} from '@/features/workflows/canvas/canvas.types';
+import type { ComputedRef, InjectionKey, Ref } from 'vue';
+import type { ExpressionLocalResolveContext } from './types/expressions';
+import type { TelemetryContext } from './types/telemetry';
+import type { WorkflowState } from './composables/useWorkflowState';
+
+export const CanvasKey = 'canvas' as unknown as InjectionKey<CanvasInjectionData>;
+export const CanvasNodeKey = 'canvasNode' as unknown as InjectionKey<CanvasNodeInjectionData>;
+export const CanvasNodeHandleKey =
+	'canvasNodeHandle' as unknown as InjectionKey<CanvasNodeHandleInjectionData>;
+export const PopOutWindowKey: InjectionKey<Ref<Window | undefined>> = Symbol('PopOutWindow');
+export const ExpressionLocalResolveContextSymbol: InjectionKey<
+	ComputedRef<ExpressionLocalResolveContext | undefined>
+> = Symbol('ExpressionLocalResolveContext');
+export const TelemetryContextSymbol: InjectionKey<TelemetryContext> = Symbol('TelemetryContext');
+export const WorkflowStateKey: InjectionKey<WorkflowState> = Symbol('WorkflowState');
