+++ packages/frontend/editor-ui/src/constants/limits.ts
@@ -1,0 +1,9 @@
+export const MAX_WORKFLOW_SIZE = 1024 * 1024 * 16; // Workflow size limit in bytes
+export const MAX_EXPECTED_REQUEST_SIZE = 2048; // Expected maximum workflow request metadata (i.e. headers) size in bytes
+export const MAX_PINNED_DATA_SIZE = 1024 * 1024 * 12; // 12 MB; Workflow pinned data size limit in bytes
+export const MAX_DISPLAY_DATA_SIZE = 1024 * 1024; // 1 MB
+export const MAX_DISPLAY_DATA_SIZE_SCHEMA_VIEW = 1024 * 1024 * 4; // 4 MB
+export const MAX_DISPLAY_ITEMS_AUTO_ALL = 250;
+
+// tags
+export const MAX_TAG_NAME_LENGTH = 24;
