+++ packages/frontend/editor-ui/src/constants/nodeTypes.ts
@@ -1,0 +1,172 @@
+import { DATA_TABLE_MODULE_NAME } from '@/features/core/dataTable/constants';
+
+export const BAMBOO_HR_NODE_TYPE = 'n8n-nodes-base.bambooHr';
+export const CALENDLY_TRIGGER_NODE_TYPE = 'n8n-nodes-base.calendlyTrigger';
+export const CODE_NODE_TYPE = 'n8n-nodes-base.code';
+export const AI_CODE_NODE_TYPE = '@n8n/n8n-nodes-langchain.code';
+export const AI_MCP_TOOL_NODE_TYPE = '@n8n/n8n-nodes-langchain.mcpClientTool';
+export const WIKIPEDIA_TOOL_NODE_TYPE = '@n8n/n8n-nodes-langchain.toolWikipedia';
+export const CRON_NODE_TYPE = 'n8n-nodes-base.cron';
+export const CLEARBIT_NODE_TYPE = 'n8n-nodes-base.clearbit';
+export const FILTER_NODE_TYPE = 'n8n-nodes-base.filter';
+export const FUNCTION_NODE_TYPE = 'n8n-nodes-base.function';
+export const GITHUB_TRIGGER_NODE_TYPE = 'n8n-nodes-base.githubTrigger';
+export const GIT_NODE_TYPE = 'n8n-nodes-base.git';
+export const GOOGLE_GMAIL_NODE_TYPE = 'n8n-nodes-base.gmail';
+export const GOOGLE_SHEETS_NODE_TYPE = 'n8n-nodes-base.googleSheets';
+export const ERROR_TRIGGER_NODE_TYPE = 'n8n-nodes-base.errorTrigger';
+export const ELASTIC_SECURITY_NODE_TYPE = 'n8n-nodes-base.elasticSecurity';
+export const EMAIL_SEND_NODE_TYPE = 'n8n-nodes-base.emailSend';
+export const EMAIL_IMAP_NODE_TYPE = 'n8n-nodes-base.emailReadImap';
+export const EXECUTE_COMMAND_NODE_TYPE = 'n8n-nodes-base.executeCommand';
+export const FORM_TRIGGER_NODE_TYPE = 'n8n-nodes-base.formTrigger';
+export const HTML_NODE_TYPE = 'n8n-nodes-base.html';
+export const HTTP_REQUEST_NODE_TYPE = 'n8n-nodes-base.httpRequest';
+export const HTTP_REQUEST_TOOL_NODE_TYPE = 'n8n-nodes-base.httpRequestTool';
+export const HUBSPOT_TRIGGER_NODE_TYPE = 'n8n-nodes-base.hubspotTrigger';
+export const IF_NODE_TYPE = 'n8n-nodes-base.if';
+export const INTERVAL_NODE_TYPE = 'n8n-nodes-base.interval';
+export const ITEM_LISTS_NODE_TYPE = 'n8n-nodes-base.itemLists';
+export const JIRA_NODE_TYPE = 'n8n-nodes-base.jira';
+export const JIRA_TRIGGER_NODE_TYPE = 'n8n-nodes-base.jiraTrigger';
+export const MICROSOFT_EXCEL_NODE_TYPE = 'n8n-nodes-base.microsoftExcel';
+export const MANUAL_TRIGGER_NODE_TYPE = 'n8n-nodes-base.manualTrigger';
+export const MANUAL_CHAT_TRIGGER_NODE_TYPE = '@n8n/n8n-nodes-langchain.manualChatTrigger';
+export const MCP_TRIGGER_NODE_TYPE = '@n8n/n8n-nodes-langchain.mcpTrigger';
+export const CHAT_TRIGGER_NODE_TYPE = '@n8n/n8n-nodes-langchain.chatTrigger';
+export const CHAT_NODE_TYPE = '@n8n/n8n-nodes-langchain.chat';
+export const AGENT_NODE_TYPE = '@n8n/n8n-nodes-langchain.agent';
+export const OPEN_AI_NODE_TYPE = '@n8n/n8n-nodes-langchain.openAi';
+export const OPEN_AI_NODE_MESSAGE_ASSISTANT_TYPE =
+	'@n8n/n8n-nodes-langchain.openAi.assistant.message';
+export const OPEN_AI_ASSISTANT_NODE_TYPE = '@n8n/n8n-nodes-langchain.openAiAssistant';
+export const BASIC_CHAIN_NODE_TYPE = '@n8n/n8n-nodes-langchain.chainLlm';
+export const QA_CHAIN_NODE_TYPE = '@n8n/n8n-nodes-langchain.chainRetrievalQa';
+export const MICROSOFT_TEAMS_NODE_TYPE = 'n8n-nodes-base.microsoftTeams';
+export const N8N_NODE_TYPE = 'n8n-nodes-base.n8n';
+export const NO_OP_NODE_TYPE = 'n8n-nodes-base.noOp';
+export const STICKY_NODE_TYPE = 'n8n-nodes-base.stickyNote';
+export const NOTION_TRIGGER_NODE_TYPE = 'n8n-nodes-base.notionTrigger';
+export const PAGERDUTY_NODE_TYPE = 'n8n-nodes-base.pagerDuty';
+export const SALESFORCE_NODE_TYPE = 'n8n-nodes-base.salesforce';
+export const SEGMENT_NODE_TYPE = 'n8n-nodes-base.segment';
+export const SET_NODE_TYPE = 'n8n-nodes-base.set';
+export const SCHEDULE_TRIGGER_NODE_TYPE = 'n8n-nodes-base.scheduleTrigger';
+export const SERVICENOW_NODE_TYPE = 'n8n-nodes-base.serviceNow';
+export const SLACK_NODE_TYPE = 'n8n-nodes-base.slack';
+export const SPREADSHEET_FILE_NODE_TYPE = 'n8n-nodes-base.spreadsheetFile';
+export const SPLIT_IN_BATCHES_NODE_TYPE = 'n8n-nodes-base.splitInBatches';
+export const START_NODE_TYPE = 'n8n-nodes-base.start';
+export const SWITCH_NODE_TYPE = 'n8n-nodes-base.switch';
+export const TELEGRAM_NODE_TYPE = 'n8n-nodes-base.telegram';
+export const THE_HIVE_TRIGGER_NODE_TYPE = 'n8n-nodes-base.theHiveTrigger';
+export const QUICKBOOKS_NODE_TYPE = 'n8n-nodes-base.quickbooks';
+export const WAIT_NODE_TYPE = 'n8n-nodes-base.wait';
+export const WEBHOOK_NODE_TYPE = 'n8n-nodes-base.webhook';
+export const WORKABLE_TRIGGER_NODE_TYPE = 'n8n-nodes-base.workableTrigger';
+export const WORKFLOW_TRIGGER_NODE_TYPE = 'n8n-nodes-base.workflowTrigger';
+export const EXECUTE_WORKFLOW_NODE_TYPE = 'n8n-nodes-base.executeWorkflow';
+export const EXECUTE_WORKFLOW_TRIGGER_NODE_TYPE = 'n8n-nodes-base.executeWorkflowTrigger';
+export const WOOCOMMERCE_TRIGGER_NODE_TYPE = 'n8n-nodes-base.wooCommerceTrigger';
+export const XERO_NODE_TYPE = 'n8n-nodes-base.xero';
+export const ZENDESK_NODE_TYPE = 'n8n-nodes-base.zendesk';
+export const ZENDESK_TRIGGER_NODE_TYPE = 'n8n-nodes-base.zendeskTrigger';
+export const DISCORD_NODE_TYPE = 'n8n-nodes-base.discord';
+export const EXTRACT_FROM_FILE_NODE_TYPE = 'n8n-nodes-base.extractFromFile';
+export const CONVERT_TO_FILE_NODE_TYPE = 'n8n-nodes-base.convertToFile';
+export const DATETIME_NODE_TYPE = 'n8n-nodes-base.dateTime';
+export const REMOVE_DUPLICATES_NODE_TYPE = 'n8n-nodes-base.removeDuplicates';
+export const SPLIT_OUT_NODE_TYPE = 'n8n-nodes-base.splitOut';
+export const LIMIT_NODE_TYPE = 'n8n-nodes-base.limit';
+export const SUMMARIZE_NODE_TYPE = 'n8n-nodes-base.summarize';
+export const AGGREGATE_NODE_TYPE = 'n8n-nodes-base.aggregate';
+export const MERGE_NODE_TYPE = 'n8n-nodes-base.merge';
+export const MARKDOWN_NODE_TYPE = 'n8n-nodes-base.markdown';
+export const XML_NODE_TYPE = 'n8n-nodes-base.xml';
+export const CRYPTO_NODE_TYPE = 'n8n-nodes-base.crypto';
+export const RSS_READ_NODE_TYPE = 'n8n-nodes-base.rssFeedRead';
+export const COMPRESSION_NODE_TYPE = 'n8n-nodes-base.compression';
+export const EDIT_IMAGE_NODE_TYPE = 'n8n-nodes-base.editImage';
+export const CHAIN_SUMMARIZATION_LANGCHAIN_NODE_TYPE =
+	'@n8n/n8n-nodes-langchain.chainSummarization';
+export const SIMULATE_NODE_TYPE = 'n8n-nodes-base.simulate';
+export const SIMULATE_TRIGGER_NODE_TYPE = 'n8n-nodes-base.simulateTrigger';
+export const AI_TRANSFORM_NODE_TYPE = 'n8n-nodes-base.aiTransform';
+export const FORM_NODE_TYPE = 'n8n-nodes-base.form';
+export const GITHUB_NODE_TYPE = 'n8n-nodes-base.github';
+export const SLACK_TRIGGER_NODE_TYPE = 'n8n-nodes-base.slackTrigger';
+export const TELEGRAM_TRIGGER_NODE_TYPE = 'n8n-nodes-base.telegramTrigger';
+export const FACEBOOK_LEAD_ADS_TRIGGER_NODE_TYPE = 'n8n-nodes-base.facebookLeadAdsTrigger';
+export const RESPOND_TO_WEBHOOK_NODE_TYPE = 'n8n-nodes-base.respondToWebhook';
+export const DATA_TABLE_NODE_TYPE = 'n8n-nodes-base.dataTable';
+export const DATA_TABLE_TOOL_NODE_TYPE = 'n8n-nodes-base.dataTableTool';
+
+export const CREDENTIAL_ONLY_NODE_PREFIX = 'n8n-creds-base';
+export const CREDENTIAL_ONLY_HTTP_NODE_VERSION = 4.1;
+
+export const EXECUTABLE_TRIGGER_NODE_TYPES = [
+	START_NODE_TYPE,
+	MANUAL_TRIGGER_NODE_TYPE,
+	SCHEDULE_TRIGGER_NODE_TYPE,
+	CRON_NODE_TYPE,
+	INTERVAL_NODE_TYPE,
+];
+
+export const NON_ACTIVATABLE_TRIGGER_NODE_TYPES = [
+	ERROR_TRIGGER_NODE_TYPE,
+	MANUAL_TRIGGER_NODE_TYPE,
+	EXECUTE_WORKFLOW_TRIGGER_NODE_TYPE,
+	MANUAL_CHAT_TRIGGER_NODE_TYPE,
+];
+
+export const DATA_TABLE_NODES = [DATA_TABLE_NODE_TYPE, DATA_TABLE_TOOL_NODE_TYPE];
+
+export const NODES_USING_CODE_NODE_EDITOR = [
+	CODE_NODE_TYPE,
+	AI_CODE_NODE_TYPE,
+	AI_TRANSFORM_NODE_TYPE,
+];
+export const MODULE_ENABLED_NODES = [
+	...DATA_TABLE_NODES.map((nodeType) => ({ nodeType, module: DATA_TABLE_MODULE_NAME })),
+];
+
+export const NODE_POSITION_CONFLICT_ALLOWLIST = [STICKY_NODE_TYPE];
+
+export const PIN_DATA_NODE_TYPES_DENYLIST = [SPLIT_IN_BATCHES_NODE_TYPE, STICKY_NODE_TYPE];
+
+export const OPEN_URL_PANEL_TRIGGER_NODE_TYPES = [
+	WEBHOOK_NODE_TYPE,
+	FORM_TRIGGER_NODE_TYPE,
+	CHAT_TRIGGER_NODE_TYPE,
+	MCP_TRIGGER_NODE_TYPE,
+];
+
+export const SINGLE_WEBHOOK_TRIGGERS = [
+	TELEGRAM_TRIGGER_NODE_TYPE,
+	SLACK_TRIGGER_NODE_TYPE,
+	FACEBOOK_LEAD_ADS_TRIGGER_NODE_TYPE,
+];
+
+export const LIST_LIKE_NODE_OPERATIONS = ['getAll', 'getMany', 'read', 'search'];
+
+export const PRODUCTION_ONLY_TRIGGER_NODE_TYPES = [CHAT_TRIGGER_NODE_TYPE];
+
+export const KEEP_AUTH_IN_NDV_FOR_NODES = [
+	HTTP_REQUEST_NODE_TYPE,
+	HTTP_REQUEST_TOOL_NODE_TYPE,
+	WEBHOOK_NODE_TYPE,
+	WAIT_NODE_TYPE,
+	DISCORD_NODE_TYPE,
+	CHAT_TRIGGER_NODE_TYPE,
+	FORM_TRIGGER_NODE_TYPE,
+];
+
+export const NODE_TYPES_EXCLUDED_FROM_OUTPUT_NAME_APPEND = [
+	FILTER_NODE_TYPE,
+	SWITCH_NODE_TYPE,
+	REMOVE_DUPLICATES_NODE_TYPE,
+	RESPOND_TO_WEBHOOK_NODE_TYPE,
+];
+
+export const NOT_DUPLICATABLE_NODE_TYPES = [FORM_TRIGGER_NODE_TYPE];
+export const UPDATE_WEBHOOK_ID_NODE_TYPES = [FORM_TRIGGER_NODE_TYPE];
