+++ packages/frontend/editor-ui/src/constants/placeholders.ts
@@ -1,0 +1,2 @@
+export const PLACEHOLDER_FILLED_AT_EXECUTION_TIME = '[filled at execution time]';
+export const IN_PROGRESS_EXECUTION_ID = '__IN_PROGRESS__';
