+++ packages/frontend/editor-ui/src/constants/samples.ts
@@ -2,6 +2,17 @@ import { NodeConnectionTypes } from 'n8n-workflow';
 import type { INodeUi } from './Interface';
 import type { WorkflowDataCreate } from '@n8n/rest-api-client/api/workflows';
 
+export const SAMPLE_PIN_DATA = [
+	{
+		name: 'First item',
+		code: 1,
+	},
+	{
+		name: 'Second item',
+		code: 2,
+	},
+];
+
 export const SAMPLE_SUBWORKFLOW_TRIGGER_ID = 'c055762a-8fe7-4141-a639-df2372f30060';
 export const SAMPLE_SUBWORKFLOW_WORKFLOW: WorkflowDataCreate = {
 	name: 'My Sub-Workflow',
