+++ packages/frontend/editor-ui/src/constants/sanitization.ts
@@ -1,0 +1,18 @@
+export const ALLOWED_HTML_ATTRIBUTES = ['href', 'name', 'target', 'title', 'class', 'id', 'style'];
+
+export const ALLOWED_HTML_TAGS = [
+	'p',
+	'strong',
+	'b',
+	'code',
+	'a',
+	'br',
+	'i',
+	'ul',
+	'li',
+	'em',
+	'small',
+	'details',
+	'summary',
+	'mark',
+];
