+++ packages/frontend/editor-ui/src/constants/selectors.ts
@@ -1,0 +1,2 @@
+export const APP_MODALS_ELEMENT_ID = 'app-modals';
+export const CODEMIRROR_TOOLTIP_CONTAINER_ELEMENT_ID = 'cm-tooltip-container';
