+++ packages/frontend/editor-ui/src/constants/templates.ts
@@ -1,0 +1,1 @@
+export const TEMPLATES_NODES_FILTER = ['n8n-nodes-base.start', 'n8n-nodes-base.respondToWebhook'];
